clients/spotify/pyproject.toml
[project]
name = "spotify-client"
version = "0.1.0"
description = "Lightweight Spotify Web API client (Client Credentials Flow)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests",
    "python-dotenv",
    "ruff",
    "black",
    "pytest",
]

# ─────────────────────────────────────────────────────────────
# Ruff configuration
# ─────────────────────────────────────────────────────────────
[tool.ruff]
line-length = 100
target-version = "py311"
fix = true
# Match isort's logic
extend-select = ["I"]      # enable import sorting checks

[tool.ruff.lint]
# Select common linting rules:
# - E/F = pycodestyle/pyflakes
# - I = import order (like isort)
select = ["E", "F", "I"]
ignore = []  # example: ignore line length if Black handles it

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false

# Optional per-file ignores
[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

# ─────────────────────────────────────────────────────────────
# Black configuration
# ─────────────────────────────────────────────────────────────
[tool.black]
line-length = 100
target-version = ["py311"]

# ─────────────────────────────────────────────────────────────
# isort configuration
# ─────────────────────────────────────────────────────────────
[tool.isort]
profile = "black"
line_length = 100
combine_as_imports = true
known_first_party = ["common", "spotify"]
# Let isort and Ruff agree
src_paths = ["src"]
order_by_type = true
force_alphabetical_sort_within_sections = true

# ─────────────────────────────────────────────────────────────
# Pytest configuration
# ─────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
pythonpath = ["../../", "."]
testpaths = ["tests"]
addopts = "-vv -s --tb=short"
